feat: Create gradle test task
Task starts the local development server, runs all connected checks,
then shutsdown the local server.

Since the app is currently running from the cloud backend, the tests
will also run on the backend. Therefore it is required to create
two versions of the asynctask - one for the debug version using
the local development server, and one for the release version using the
cloud backend.

Also, as the paid free version now includes an interstitial add the
existing tests will only work on the paid version so they need to
run in that flavor only.